prometheus_exporter/requirements.txt
prometheus-client==0.7.1
psutil==5.9.0
paho-mqtt==1.6.1